integration tests environment
Use https://github.com/puppetlabs/rspec-system-puppet to implement
vagrant based integration tests. The README.md is updated with
instructions to run the integration tests locally. These tests could be
run by a deamon listening to the gerrit stream, when a changeset is
submitted for review, to prove that they do not introduce a regression.

The dependencies from .fixtures.yml have been manually replicated to
spec/spec_helper_system.rb . rspec system puppet is still young and
people do this. It will be a burden only for long term maintenance. And
when it does, in a few months from now, it is safe to assume that rspec
system puppet will have support for reading from an existing file. The
alternative is that rspec system puppet is obsoleted / dead, in which
case this is even a bigger problem. In any case, if work had to be done
to fix this, it should be a patch against rspec system puppet and not
part of puppet-ceph.
https://github.com/puppetlabs/rspec-system-puppet/issues/5

A test for the ceph::repo class is included to demonstrate the
environment actually works. It is not meant to fully test the ceph::repo
integration this would require a separate patch. rspec-system-puppet
requires at least one file in spec/system otherwise it will fail with
an unrelated error:

   Could not autoload puppet/provider/ceph_config/ini_setting: undefined
   method `provider' for nil:NilClass (Puppet::Error)

This test can be removed as soon as another is added.

Change-Id: Ib65120a58ce203126be5897f602e280c474642dc
Signed-off-by: Loic Dachary <loic@dachary.org>