Make ceph::repo ensurable + add support for emperor
The ensure argument is added to ceph::repo and used in integration tests
to make sure the repository is removed the repository before installing
it. The integration tests do not re-create a virtual machine for each
test and the leftovers of the previous test must be cleaned.

The emperor Ceph release is now supported to be the default.

Closes-Bug: #1255940

Change-Id: Ib6da26b95be9f549d55bd88f867b43d6aa536f10
Signed-off-by: Loic Dachary <loic@dachary.org>