remove rspec-system dependencies from Gemfile
rspec-system will fail to install on jenkins slaves. It does not matter
anyway since it is not going to be used for unit tests. Create a
Gemfile-rspec-system to be used when running integration tests. Although
it would be convenient to be able to

   BUNDLE_GEMFILE=Gemfile-rspec-system bundle exec rake spec:system

it will fail because the environment variable won't be propagated where
it should.

   .rspec_system/vagrant_projects/x64/Gemfile-rspec-system not found

The README.md file is updated to simply mv to Gemfile which is good
enough for the environment running the integration tests.

A bug was filed against openstack-ci to resolve this installation
problem but it is unclear if and how it could be fixed
https://bugs.launchpad.net/openstack-ci/+bug/1290710

Change-Id: I0bff624892837536419bdbdd6651b009600fe9c6
Signed-off-by: Loic Dachary <loic@dachary.org>