ceph::osd : add support for mon data directories
The creation of the OSD is separated in two stages, ceph-disk prepare
and ceph-disk activate. When dealing with a device this is not
necessary. But it is necessary when mon data is a directory.

The mon data directory is created if it does not exist. Although this is
not necessary for versions >= dumpling, it is for cuttlefish.

The tests are unmodified : they are run once with the /dev/sdb device
and again with the /srv/data directory.

The filestore_xattr_use_omap = true is added because cuttlefish does not
set it by default and will fail if trying to OSD::mount a directory from
a file system with limited xattr capacity such as ext4.

Note: the indentation in the tests is intentionally incorrect to not
introduce white space changes and make a clean diff.

Change-Id: Ie40eea0908b0f2e14420e2b0df2ba26e1465d23e
Signed-off-by: Loic Dachary <loic@dachary.org>