Support for RedHat in ceph::mon
Support for RedHat in ceph::mon and in ceph::mon, ceph::keys CI tests.

Yumrepo isn't propperly removed but only disabled. This leads to a need
to ensure => absent on the ceph::repo whenever a switch to a different
ceph version happens. Therefore it's refactored run tests ordered by
release

Change-Id: I51150c1e3316bdab20c4458e79c4aeeb15fe4bd2
Closes-Bug: #1264916
Closes-Bug: #1323841