Create rspec-system tests for RedHat family
Provides the necessary nodset configurations and examples
on how to run the tests on RedHat family systems.

Using CentOS 6.4 as a representative for RedHat family systems,
as there is a box available from the puppetlabs Vagrant boxes
repository with the same Virtualbox version requirements as the
existing Ubuntu Precise one:

http://puppet-vagrant-boxes.puppetlabs.com

This is needed to test solutions for bugs 1264959 and 1264916.

There is a circular dependency issue between those bugs and this:
Cannot test the solutions to those bugs without this patch, but without
solutions to those tests the rspec-system tests for RedHat family
systems won't run. To have only running code committed, this should
only be merged together with the other patches.

Change-Id: I2093635034ca78b7fb6dbfd950cf35af8a03bc6f
Closes-Bug: 1253211
Related-Bug: 1264959
Related-Bug: 1264916