Fixed removal command in ceph::osd
And updated the CI tests. ceph::repo needs to be removed whenever
a different release is targeted. This is mainly because "enabled" is
used in yumrepos, because of the lack of "ensure" in Puppet < 3.5
Refactored to not have to switch ceph::repo constantly, but run tests
on a per release order.

Change-Id: I555acda7413765aaa99f922893649e1f13215450
Closes-Bug: #1264916
Closes-Bug: #1323841
Closes-Bug: #1323838