Remove unnecessary if statement in ceph::mon
A previous patchset implementing RedHat family support for ceph::mon had
an erroneous if statement in the OS selection code.
This is fixed here, in its own patchset though, as changes to the previous
patchset failed the CI tests (and even crashed the test infrastructure as
it got stuck indefinitely.)

The removal of the if statement requires some changes in the tests as well.

Also added a check for the exec in RedHat family.

Change-Id: I38fbf189c806a21c40bb55dd47d6debb3a592b89