obsolete cuttlefish and add firefly
Firefly is the latest stable Ceph released and was published May
2014. It is added to all integration tests. The cuttlefish release is
obsoleted to reduce the burden of backward compatibility down to the
last three stable releases.

The journal size for tests is made to be 100MB to use less disk during
testing and lower the risks of running of out space.

Change-Id: I35ad5d702a5f8094bb72c67b87582154e2c017fc
Signed-off-by: Loic Dachary <loic@dachary.org>