ceph::osd support for scenario_node_terminus
scenario_node_terminus allow users to specify a data layer that lives on
top of Puppet and allows an additional level of flexibility to
deployments.
It was designed to compose a set of core Puppet modules into multiple
deployment permutations (or reference architectures) called scenarios
while requiring as little duplicate code as possible.

The above description is taken from the source:
https://github.com/bodepd/scenario_node_terminus/

This patch adds the necessary dependency information to manage
ceph::osd instances while using scenario_node_terminus.

An integration test which can serve as an example configuration is also
provided.

Change-Id: I1bb2b402028df66478039535a3c7b48cbd05dcbb