Support for scenario_node_terminus for ceph::mon
scenario_node_terminus allow users to specify a data layer that lives on
top of Puppet and allows an additional level of flexibility to
deployments.
Is was designed to compose a set of core Puppet modules into multiple
deployment permutations (or reference architectures) called scenarios
while requiring as little duplicate code as possible.

The above description is taken from the source:
https://github.com/bodepd/scenario_node_terminus/

This patch adds the necessary dependency information to support
scenario_node_terminus:

* Configuring the repositories before installing the package
* Configuring /etc/ceph/ceph.conf before adding a monitor

The last needed dependency of installing the package before configuring
/etc/ceph/ceph.conf was already in place, but is necessary for this as
well.

To show the dependency information is correct, there is a basic
integration test which installs one monitor without cephx on one host
using scenario_node_terminus.

Change-Id: Ie160c8f644d7324177b5ced00e7d5e75257e359b