Initial classes/tests for a roles/profiles pattern
This is a first step to providing roles/profiles manifests according
to the roles/profiles pattern.

The roles/profiles pattern is a way to combine ceph manifests into
functional units:

Profiles combine multiple manifest to provide a single service. For
example a ceph monitor server needs the repository, packages,
configuration and finally the monitor service.

Roles define sets of profiles to configure a specific server. For
example an allinone role would install the monitor as well as an osd
profile.

The learn more have a look at:
http://www.slideshare.net/PuppetLabs/roles-talk

To start with the basic configuration this provides:
* params: extracts the configuration from hiera
* base: installs ceph and configures ceph.conf

All configuration happens via hiera. Example hiera files are
provided.

NB: this uses hiera autoloading and will only work with Puppet >=3.0

Change-Id: Iba9aca7f124bd3a719dc18292b18fc0f4f386d5c