Mon profile for roles/profiles pattern
This provides a profile for monitor instances in the context of a
roles/profiles pattern.

The roles/profiles pattern is a way to combine ceph manifests into
functional units:

Profiles combine multiple manifest to provide a single service. For
example a ceph monitor server needs the repository, packages,
configuration and finally the monitor service.

Roles define sets of profiles to configure a specific server. For
example an allinone role would install the monitor as well as an osd
profile.

The learn more have a look at:
http://www.slideshare.net/PuppetLabs/roles-talk

All configuration happens via hiera. Example hiera files are
provided.

Change-Id: If07d18cdb202706c8478570027e80a85c8159d31
NB: this uses hiera autoloading and will only work with Puppet >=3.0