Rewrite data model to be pure hiera
This patch adds experimental support for a distribution that uses pure
hiera in place of the scenario_node_terminus.

Data mappings are expressed using interpolation except
in the case where they are non-string, in which case
yaml anchors are used. At this time, YAML anchors are not
working across files, which places some restrictions on
their use.

Scenarios are now a part of the hiera_data folder,
and each scenario yaml file contains its mappings,
globals, and in future probably its 'user.scenario.yaml'
data.

Roles are under scenarios/%scenario/$role, and use a hiera
lookup of 'classes' and 'class_groups' to include classes.

Class groups are a new yaml file in hiera_data, and have
been changed to be string quoted since starting yaml keys
with a % is invalid.

The original idea as well as the above description for this data model
is by Micheal Chapman for stackforge/puppet_openstack_builder.

Change-Id: I56203339cf28bf25f7518772596d0175b064499e