Update CI boxes
Update the CentOS 6 CI box from 6.4 to 6.5. This is needed to support
RDO which in turn is a requirement for the rgw/keystone CI tests.

Add CentOS 7 and Ubuntu 14.04 boxes so we can CI test them as a basis
to add support for them.

Because rspec-system and rspec-system-puppet are deprecated in favor
of beaker add dependencies on a fork to provide the necessary features.
This is temporary until all tests are switched to beaker.

Change-Id: I0ef014ff73a431a691fc25c35f6461fec614c337