Increase flexibility with managing client keys in profile
This commit introduces the client_keys hash parameter to
ceph::profile. This allows one to use the profile to inject
and configure any client keys with any parameter.

Backwards compatibility is *NOT* maintained for the purpose
of staying as clean and as simple as possible in the context
of drafting the first stable release of puppet-ceph.
This avoids shipping the first stable release with, already,
a deprecation that will have to be maintained throughout an
entire release cycle.

Some tests have been added and test coverage is in general
improved as part of this commit.

Change-Id: Ie6adbd601388ab52c37037004bd0ceef9fc41942