Fix idempotency when bootstrapping ceph::mon with cephx
During bootstrapping with cephx ceph::mon needs a temporary file
under /tmp/ to store the mon. key. This gets laters removed again
due to security reasons.

This changes the creation of the key from a file to an exec with
an unless clause so the key doesn't get recrreated and removed
during every subsequent run.

Change-Id: Ied38afc03d68f7868b1d1320752a48989fa2d018