acceptance: fix ceph_rgw_civet_spec tests
In ceph_rgw_civet_spec, we used to deploy fedorapeople repo, which does
not exist anymore, since it moved to centos7.
To unblock the gate, this patch aims to update URLs & GPGkeys so it wil
work again.

Though in the future, we will move all those tests in
puppet-openstack-integration and use common bits that we already have in
place, also deploy current master and not old versions of OpenStack.

Change-Id: I64e69c9048523a8bb3d0852d2d642df677b80ab5