Ensure list of packages to install is user customizable
There are hardcoded references to the 'ceph' package in the module but
the package isn't provided by the ceph.com repos anymore. This change
fixes the dependencies by referring them to ::ceph::params::packages instead

It also changes the type of the packages parameter to an array so that
multiple packages can be passed.

Change-Id: I13462219522386f8740b0d70916a44f3474115e4
Closes-Bug: 1630468