Update to reduce puppet errors.
This patch reduces the amount of warnings due to 'unknown variables'
with certian combinations of class arguments. mon_caps, osd_caps and
mds_caps are set to '' if no relevant options are passed.
cluster_option is set to '--cluster ceph' if no custom cluster name
is passed to the relevant classes.

Change-Id: I3204b402c00dab4836c8a3ea64eb769a83f7b51f