Fail deployment if non-existent block devices are passed for OSDs
If a user passes a block device, e.g. /dev/sdx, that doesn't actually
exist, then the user made a mistake and the deployment should fail.
Without this change if /dev/sdx doesn't exist, then a directory is
created called /dev/sdx and set up as a directory-backed OSD. This
patch adds a check to fail the deployment if the requested OSD is
a directory with a name starting with /dev.

Change-Id: I5f03fc332b8df3914dc8609a203eb38693cb87a7
Closes-Bug: 1665697