Some fixes for the ceph disk activate part
The "unless" condition for the ceph disk activate part did not work when using block devices as puppet always tried to activate a running OSD again and again. So I introduced an additional check with "ceph-disk list" which checks if the OSD is already active. The test to check if a whole block device gets used was not working for me as well. Udev should activate a whole block device when running ceph (since Infernalis), but for me the module tried to activate a OSD.