fix the service names
The name of the service which puppet should handle is based on the "name" directive of the service definition. For an upstart managed job this should be "ceph-mon" as otherwise the start script can not be found. Also the service_provider in the params class is called "Debian" for Debian based distributions which does not work. We should use the actual service_provider from the system ("upstart" or "systemd").