Update ceph-osd-check-udev to correct an awk condition
The awk statement will print 1 if the ceph release it < 0.94.
Otherwise, it will print nothing. When it does not print anything,
no value is set to DISABLE_UDEV. This results in the test failing.

++ ceph --version
++ awk 'match($3, /[0-9]+\.[0-9]+/) {if (substr($3, RSTART, RLENGTH) <= 0.94) {print 1}}'
+ DISABLE_UDEV=
+ test -f /usr/lib/udev/rules.d/95-ceph-osd.rules
+ test -eq 1
sh: line 3: test: -eq: unary operator expected

This patch adds a 'else { print 0 }' if the substr does not return true.

++ ceph --version
++ awk 'match($3, /[0-9]+\.[0-9]+/) {if (substr($3, RSTART, RLENGTH) <= 0.94) {print 1} else { print 0 } }'
+ DISABLE_UDEV=0
+ test -f /usr/lib/udev/rules.d/95-ceph-osd.rules
+ test 0 -eq 1

Change-Id: I4d2da48717c57b0d8d853d9e2a711ca02b18c6e1