Fix lspool issues on Mimic
With the Mimic release they have changed
the output of the "ceph osd lspools" command.

The old output (from Luminous):
$ceph osd lspools
5 glance,6 gnocchi,13 cinder,25 nova,

The new output (from Mimic):
$ceph osd lspools
1 glance
2 nova
3 cinder
4 gnocchi

This changes the usage of "ceph osd lspools"
to "ceph osd pool ls" which exists in both
versions and has the same output.

Example output of "ceph osd pool ls"
$ceph osd pool ls
glance
nova
cinder
gnocchi

We combine this with a grep with word-regexp
which will make sure we do a fullmatch on the
name of the pools.

Change-Id: I8883c8400ffc18672479fe2ad8cc7955a52bc466