Do not check for ldap secret with Fedora and RedHat > 7
Ceph rgw fails when running with Fedora as it checks
for not existing file: /etc/openldap/secret which is
only required when enabling ldap [1].

This patch explicitly set rgw_ldap_secret to null to
skip ldap_secret check. This can be cleared up once
issue[1] is fixed in ceph luminous.

[1] https://tracker.ceph.com/issues/24228

Change-Id: Icb5d5b5f3c5348fda00035189591f006bf2fac5d