Added support for the rgw_swift_url parameter in the rgw.pp
This parameter is needed for the swift API if there is a load balancer
in front of the rados gateway servers. If not provided and the swift
access url is different than http://$::fqdn:port, the swift API is not
working (users get a connection error).

http://docs.ceph.com/docs/mimic/radosgw/config-ref/

Moreover, use the rgw_dns_name option both for civetweb and apache.
This is a valid option for both frontend types, and becomes necessary
if it has been changed at least once in the past.

Change-Id: I639d15e30b0ba6c94614c1650d4ea85258ae6ede