make rgw_dns_name be truly optional for multi-hostname configs
currently setting it to `false` sets domain to "false" which is not usefil
make it so `false` removes setting from the config.

This is needed for example when setting up zonegroup to work on more than one domain (the `hostnames`) setting,
in that case rgw_dns_name must be unset for rgw to work with set up domains.