Prepare the final release of stable/jewel
Jewel is quite old and is no longer used in our CI. To retire the old
branch, the final release will be created.

Because this version has several new features since the last release,
ideally we should bump minor version. However unfortunately 2.5.0 is
already assigned, so this bumps the patch version instead.

Change-Id: Ic17d324b975ae4fd1649f82c2fd26dd08b4e1768