rgw profile: Support keystone integration
This introduces a few new parameters to the profile class so that users
can enable keystone integration of RADOS Gateway.

This allows us to fix some ignored profile parameters, and also remove
the references from the non-profile manifest to the profile parameters.

The rgw_keystone_version parameter is removed by this change.
The actual parameter was deprecated[1] and was removed[2] from the rgw
class very long ago.

[1] 85b9d61c40430de32d311cea201a2a33ec91a6b8
[2] 0377da4e088ab88b97af1143191be8da39e1623b

Change-Id: I4026d3c2d40ae9b7ed9c3a60529011854b878f89