Remove CentOS/RHEL 7 support
The train release was the last release with CentOS 7 support, but it is
now being transitioned to EOL.

This change removes support for CentOS 7 and RHEL 7 because we no
longer support any OpenStack release supporting these old operating
systems.

Because fastcgi package was available for only CentOS/RHEL 7, its
support is also removed by this change.

This also removes the wrong warning message for rgw deployment with
mod_proxy_fcgi. This pattern is still valid in supported operating
systems and only the one with mod_fastcgi should be deprecated or
removed.

Change-Id: I1238b23f1aee57641788a21556123bec94a1ed1b